+++ stl/lib/stl.dart
@@ -6,3 +6,5 @@ library;
 export 'src/collections/vector.dart';
 export 'src/collections/deque.dart';
 export 'src/collections/stack.dart';
+export 'src/collections/forward_list.dart';
+